+++ CHANGELOG.md
@@ -11,6 +11,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 
 ### Added
+* $-expression FMT formatting in specification file
+* User-defined datatypes in specification file
+* Attribute support for data type
 
 ### Changed